requirement.txt
flake8
flake8-docstrings
pep8
pyflakes
pydocstyle
flask>=1.0.2
flask-swagger==0.2.14
pytz>=2015.4
cerberus==1.2
PyYAML==5.1
gunicorn==19.7.1
honcho==1.0.1
python3-ldap
